# perf: improve input trimming