--- a/src/checkers.cpp
+++ b/src/checkers.cpp
@@ -79,4 +79,16 @@ namespace ada::checkers {
     return std::all_of(iterator_start, iterator_end, checkers::is_digit);
   }
 
+  ada_really_inline constexpr bool is_next_equals(const std::string_view::iterator start,
+                                              const std::string_view::iterator end,
+                                              const char c) {
+    return (std::distance(start, end) > 0) && (start[1] == c);
+  }
+
+  ada_really_inline constexpr bool is_not_next_equals(const std::string_view::iterator start,
+                                              const std::string_view::iterator end,
+                                              const char c) {
+    return (std::distance(start, end) > 0) && (start[1] != c);
+  }
+
 } // namespace ada::checkers