Before:

```
BasicBench_AdaURL            8234 ns         8233 ns        84999 time/byte=10.1898ns
```

After:

```
BasicBench_AdaURL            8075 ns         8075 ns        86405 time/byte=9.99371ns
```